Add files via upload
Grayscale images of 15 subjects under 10 different conditions were obtained and are
given in the file Dataset_Question1. Due to storage limitations, only one representative image
can be stored for each subject in the database for future automated facial recognition
purposes. Perform SVD on all the images for a subject to identify characteristic features that
will be stored for each subject. Only this information should be used in the future automated
facial recognition tasks. In the report, please show the representative images of all the
subjects. Given an image, the facial recognition method is based on the smallest norm
between the image and the representative images in the database. Determine the number of
images out of 150 that you are able to correctly identify based on this approach in terms of
accuracy. MATLAB image processing functions or python image processing libraries are not
to be used for this assignment